requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.10
tabulate>=0.9
matplotlib>=3.7
rich>=13.0        # monitor_live.py 终端彩色显示
pyzmq>=25.0       # data_bus ZMQ 行情广播
pyarrow>=14.0     # data_bus Parquet 分片写入与读取
psutil>=5.9       # process_watcher.py 进程检测
fastapi>=0.112    # Web 控制台 API
uvicorn>=0.30     # Web 控制台服务启动
